Finished the apis

+++ todo-backend/src/main/java/todobackend/todo/service/TodoService.java
@@ -3,7 +3,9 @@ package todobackend.todo.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import todobackend.todo.model.Todo;
+import todobackend.todo.model.User;
 import todobackend.todo.repository.TodoRepo;
+import todobackend.todo.repository.UserRepo;
 
 import java.util.List;
 
@@ -13,13 +15,17 @@ public class TodoService {
     @Autowired
     private TodoRepo todoRepo;
 
+    @Autowired
+    private UserRepo userRepo;
+
     public List<Todo> getAllTodos() {
         return todoRepo.findAll();
     }
 
-    public Todo getTodo(int todoid) {
-        // TODO
-        return null;
+    public List<Todo> getAllUserTodo(int userid) {
+        User user = userRepo.findUserByUserid(userid);
+        if(user == null) return null;
+        return user.getTodos();
     }
 
     public Todo createTodo(Todo todo) {     // Returnerer denne Todo object?
@@ -27,12 +33,20 @@ public class TodoService {
     }
 
     public Todo updateTodo(int todoid, Todo updatedTodo) {
-        // TODO
-        return null;
+        Todo todo = todoRepo.findByTodoid(todoid);
+        if(todo == null) return null;
+
+        todo.setTodoheader(updatedTodo.getTodoheader());
+        todo.setTodoinfo(updatedTodo.getTodoinfo());
+        todoRepo.save(todo);
+        return todo;
     }
 
-    public void deleteTodo(int todoid) {
+    public Todo deleteTodo(int todoid) {
+        Todo todo = todoRepo.findByTodoid(todoid);
+        if(todo == null) return null;
         todoRepo.deleteById(todoid);
+        return todo;
     }
 
 }
